WIP: Initial trip creation implementation
The Trips page now displays a Create Trip button, which brings the user
to the /createtrip page. This commit also includes the Google Maps
Places API, which we use in the Create Trip page to select the
destination for the trip. Still need to save the destination information
to the app state, and use it to call the backend API to create the trip.
[#129431461]