Avoid mutating reducer arguments
Previous implementations of the reducers would perform operations that
mutated the state object before passing it back. For example, some
operations deleted properties of the object. However, react-redux
strictly forbids mutating the object, as discussed here:
http://redux.js.org/docs/Troubleshooting.html#never-mutate-reducer-arguments

One of the side effects of doing it incorrectly was that
mapStateToProps() wasn't triggered for one of the containers, leading to
an issue where a page's props weren't being refreshed when expected to.

This commit fixes the issue for all known violations.
[Finishes #129559315]