Support adding ideas to trip on Web
A search box at the top of the Ideas list allows users to add ideas. The
box queries the Google Maps Places Autocomplete API. Once a place is
selected and the user presses Add, the app creates the trip idea on the
backend and displays the updated list of ideas on the page.

[Finishes #129293927]