Create new ErrorMessage component
Multiple pages share the same error message HTML code, so it makes sense
to spin off into a standalone component.

Also - since it's more beneficial for styles to be defined within a
component (so that styles can be dynamically generated based on props),
this commit will start a movement towards moving style definitions back
to the components. [#130952719]