Keep grey placeholder box displayed until another idea selected
When dragging an idea, the grey placeholder box for where the idea's
supposed to be will not disappear until the user hovers over another
idea in the list. At that point, that idea will have the placeholder box
displayed in its place instead.

NOTE: There's an unsolved edge case where if you start dragging from a
part of the UI that is outside the drop target, the grey placeholder box
does not display. Will fix this bug in a future commit.

[#130887357]