Change approach for trip idea reorder implementation
The previous approach kept all the trip idea UI elements in place, and
intended to change the content of each element based on the destination
position for the dragged idea. That approach was too cumbersome and
required maintaining too much state.

This new approach triggers a redux action to reorder ideas for each new
position the currently dragged idea hovers over. It won't save the idea
permanently until after the drag-and-drop ends and there's been a change
in the order of the ideas array. [#130887357]