UI complete for trip idea reordering
The new approach works! The only suboptimal solution being the
modification of the drag source's item object to keep the index up to
date.

With this change, you can now drag a trip idea to another location in
the list, let go, and the array would visually appear updated.

The next commit will hook this up with the backend API to persist state.
[#130887357]