Don't replace trips state after deleting trip
The backend API doesn't return the list of trips upon deleting a trip,
so we won't replace the trips object in the state. This fixes a bug
where deleting the trip causes the trips object to be undefined,
therefore breaking navigation back to the Home page from a trip page.
[Finishes #131149197]