Dynamically determine the map's width
The map element had a width set to 100% and a padding-left value.
This configuration caused problems where many of the map's elements and
UI were rendered at an offset equal to the padding-left value.

To address this issue, we change the offset to a positioning offset (by
setting the `left` CSS property) and dynamically determine with the
width of the map element using the window's innerWidth and the fixed
width of the left section. [Finishes #131329129]