Ensure map fits bounds of all the ideas' locations
The default viewport for the map is the trip destination's viewport. But
if the user adds ideas that extend beyond the viewport, the map will now
zoom out to ensure all the ideas fit within the bounds. We add a padding
of 100px so that no location lies on the edge of the map.

[#131305007]