Only save new idea to state if it's a valid Google place
The Google Autocomplete API exposes a getPlace() API that fetches
information about the currently selected place. However, it gives no
guarantee that there is a place selected; in other words, if the user
types random characters on the input box without selecting a place from
the dropdown, that API returns an object with the `name` field set to
the input text.

This commit adds a check to ensure that we're saving a real google Place
object before we save it to state. [Finishes #131923975]