Lock left column scroll while autocomplete dropdown is visible
Because the window's scrolling behavior is locked (so that only the left
column scrolls, and not the map on the right), the autocomplete dropdown
doesn't know how to scroll with the left column while it is visible.

To remedy this issue, we will programmatically lock the left column from
scrolling while the autocomplete dropdown is visible.

[Finishes #131909473]