Only generate dropdown menu component if user is logged in
The web app was crashing on unauthenticated loads because the render()
method was creating the dropdown component and assigning it to a
constant variable, and it required access to the user's name, which
obviously doesn't exist if no user is logged in yet.

To fix this, wrap the generation of the component behind a function so
that it's only called when a user object exists in the redux store.
[Finishes #131950463]