Display markers and popups using Mapbox layers and sources
Layers and sources are a much better avenue of visualizing a series of
data on a Mapbox map. They support dynamically drawing markers and
popups on the canvas element (as opposed to using HTML & CSS).

Going with this approach allows us to avoid needing to keep track of the
list of markers in redux state, and enables us to draw invisible layers
that can be used as hover targets to display popups so that they don't
disappear once the mouse leaves the visible markers.

[Finishes #132045503]