Use flexbox layout to fit long trip titles
If a trip title wraps to multiple lines, the text will appear outside of
the trip entry's box on the trips page. To fix this, display the map and
the title in a div with flex layout. This would shrink the map
container's height as necessary so that the title is visible and stays
within the box. [#130872535]