Redirect upon successful login/signup if specified
Use case: I'm viewing a trip page in guest mode. I press the Login
button, and I see the login page. After I log in (or create an account),
I will be redirected back to that trip page.

This commit structures the logic so that it can support any kind of
redirect - the redux store stores a callback function for the redirect.

[Finishes #130872519]