Override error message for unauthorized trip views
Displaying a "Not Authorized" error message when someone visits a trip
page implies that a trip exists, which could affect the trip owner's
perceived privacy. To remedy this issue, we will override that error
message with "Trip not found" so that anyone who sees this trip can't
distinguish whether the trip exists or they don't have permissions.

[#133411311]