Only show error message in one place on trip page
When the trip page and a modal are open at the same time and an error
occurs, both the modal and the trip page would display the error
message, since both UIs react to the tripState error string. To avoid
this duplication, we will not show any error message on the Trip page if
a modal is open. [Finishes #133421961]